--- src/subdomains/core/payment-link/share/c2b-providers.enum.ts
@@ -1,11 +1,0 @@
-import { Blockchain } from '../../../../integration/blockchain/shared/enums/blockchain.enum';
-
-export enum C2BPaymentProvider {
-  BINANCE_PAY = Blockchain.BINANCE_PAY,
-}
-
-export const C2BPaymentLinkProvider = {
-  BINANCE_PAY: C2BPaymentProvider.BINANCE_PAY,
-} as const;
-
-export type C2BPaymentLinkProvider = (typeof C2BPaymentProvider)[keyof typeof C2BPaymentProvider];
